BitStream: rewrite BitStreamCacheRightInLeftOut.
Previously, the bits within BitStreamCacheRightInLeftOut were layed out like:
63|                        | fill level|0
  |consumed but not flushed|active bits|0
So to extract the first two bits ("ac"), you'd need to both shift
(producing "|000000consumed but not flushed|ac"),
and then mask the bits you want to extract.

This is rather convoluted. E.g. on pre-BMI x86 cpu's
that's ~6 instructions, on BMI1 cpus that's 4 instructions
(shift+mov+or + BEXTR), on BMI2 just two (shift+BZHI).

But if we change the cache layout so that the active are at the high end,
with no garbage ontop, then we only need a single shift instruction
to extract the bits (well, and `64-count`, which is just `neg` for
x86/aarch64, to calculate the shift amount), in any case.

The downside is of course that `push()` needs to do 2 more `sub`..

The perf impact will obviously radically differ, but on
BMI1 machine this seems to regress Nikon and Phase One decompressors...