Rewrite BitStream. Don't inherit from ByteStream. Parametrize the replenisher
The main idea here is that the coupling to ByteStream is too strong.
And that coupling is kinda just wrong, we don't care which byte order
the buffer has. And because of that coupling we can't change
how we actually consume the input buffer.

So lets just decouple them.

The obvious effect is that 'cache::cache' now resides
at the very beginning of the struct, which is is not bad,
although shouldn't matter because ByteStream should be getting SROA'd,
if it isn't and remains on stack, the perf suffers anyways.

And BitStream slightly decreases in size, but doesn't matter really.

This has no perf impact of note on the masterset, changes are in the noise.